clearer error messages and use of raise (#36)
* clearer error messages and use of raise

This is an attempt at improving on the error messages. Before this update when a required parameter is missing a simple print tells you about the error but does not stop the program. hence the program only stops when the parameter is required and not readily available. The print about the missing parameter is more or less cryptic and the actual error of python make this very confusing.
This changes this behaviour. When a parameter is missing the a raise ValueError is used with a more precise message on what is missing. Using the raise forces the program to stop while parsing the parameter file.

* Fix typo